+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Added
 - Officially supported docker release
+- Ability to remove followers unilaterally without a block
 
 ## Changes
 - Follows no longer override domain blocks, a domain block is final